Refactor NavBar component by extracting hooks and utils
